refactor: extract client-types.ts from client.ts (1,848 lines)
Move all exported interfaces, enums, type aliases, and the ApiError class
from client.ts into client-types.ts. client.ts re-exports everything via
`export * from "./client-types"` so no consumer imports change.

client.ts: 6,062 → 4,392 lines (28% reduction)
client-types.ts: 1,848 lines (pure types, zero runtime)

Method splitting (prototype augmentation) was evaluated but deferred —
the remaining 4.4K lines are a single class where ~270 methods call
shared private infrastructure (fetch, rawRequest). Splitting via
declaration merging adds TypeScript complexity for diminishing returns.
The types extraction was the high-value, zero-risk win.
